backend/requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
faiss-cpu
langgraph
langchain
elevenlabs
google-generativeai
python-dotenv
